# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountDeleteRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "maxLength": 72,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "x-rentivo-max-utf8-bytes": 72
          }
        },
        "required": [
          "password"
        ],
        "title": "AccountDeleteRequest",
        "type": "object"
      },
...
      "PasskeyRegistrationCompleteRequest": {
        "additionalProperties": false,
        "properties": {
          "challenge_id": {
            "minLength": 1,
            "title": "Challenge Id",
            "type": "string"
          },
          "credential": {
            "$ref": "#/components/schemas/WebAuthnRegistrationCredential"
          },
          "name": {
            "default": "Minha Passkey",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "challenge_id",
          "credential"
        ],
        "title": "PasskeyRegistrationCompleteRequest",
        "type": "object"
      },
      "PasskeyResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "uuid": {
            "title": "Uuid",
            "type": "string"
          }
        },
        "required": [
          "uuid",
          "name",
          "created_at"
        ],
        "title": "PasskeyResponse",
        "type": "object"
      },
      "PasswordChangeRequest": {
        "additionalProperties": false,
        "properties": {
          "confirm_password": {
            "maxLength": 72,
            "minLength": 1,
            "title": "Confirm Password",
            "type": "string",
            "x-rentivo-max-utf8-bytes": 72
          },
          "current_password": {
            "maxLength": 72,
            "minLength": 1,
            "title": "Current Password",
            "type": "string",
            "x-rentivo-max-utf8-bytes": 72
          },
          "new_password": {
            "maxLength": 72,
            "minLength": 1,
            "title": "New Password",
            "type": "string",
            "x-rentivo-max-utf8-bytes": 72
          }
        },
        "required": [
          "current_password",
          "new_password",
          "confirm_password"
        ],
        "title": "PasswordChangeRequest",
        "type": "object"
      },
...
      "PixUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "pix_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Pix Key"
          },
          "pix_merchant_city": {
            "anyOf": [
              {
                "maxLength": 15,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Pix Merchant City"
          },
          "pix_merchant_name": {
            "anyOf": [
              {
                "maxLength": 25,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Pix Merchant Name"
          }
        },
        "title": "PixUpdateRequest",
        "type": "object"
      },
      "PixUpdateResponse": {
        "additionalProperties": false,
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/ProfileResponse"
          }
        },
        "required": [
          "profile"
        ],
        "title": "PixUpdateResponse",
        "type": "object"
      },
      "Problem": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Fields",
            "type": "object"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "request_id"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "RecoveryCodesResponse": {
        "additionalProperties": false,
        "properties": {
          "recovery_codes": {
            "items": {
              "type": "string"
            },
            "title": "Recovery Codes",
            "type": "array"
          }
        },
        "required": [
          "recovery_codes"
        ],
        "title": "RecoveryCodesResponse",
        "type": "object"
      },
...
      "TOTPConfirmRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "TOTPConfirmRequest",
        "type": "object"
      },
      "TOTPDisableRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "maxLength": 72,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "x-rentivo-max-utf8-bytes": 72
          }
        },
        "required": [
          "password"
        ],
        "title": "TOTPDisableRequest",
        "type": "object"
      },
...
      "WebAuthnRegistrationCredential": {
        "additionalProperties": false,
        "properties": {
          "authenticatorAttachment": {
            "anyOf": [
              {
                "enum": [
                  "cross-platform",
                  "platform"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authenticatorattachment"
          },
          "clientExtensionResults": {
            "$ref": "#/components/schemas/WebAuthnRegistrationExtensions"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "rawId": {
            "minLength": 1,
            "title": "Rawid",
            "type": "string"
          },
          "response": {
            "$ref": "#/components/schemas/WebAuthnAuthenticatorAttestationResponse"
          },
          "type": {
            "const": "public-key",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "rawId",
          "type",
          "response"
        ],
        "title": "WebAuthnRegistrationCredential",
        "type": "object"
      },
      "WebAuthnRegistrationExtensions": {
        "additionalProperties": false,
        "properties": {
          "credProps": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WebAuthnCredentialProperties"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "WebAuthnRegistrationExtensions",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/security/change-password": {
      "post": {
        "operationId": "change_password_api_v1_security_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation problem"
          }
        },
        "summary": "Change Password",
        "tags": [
          "security"
        ]
      }
    },
    "/api/v1/security/delete-account": {
      "post": {
        "operationId": "delete_account_api_v1_security_delete_account_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation problem"
          }
        },
        "summary": "Delete Account",
        "tags": [
          "security"
        ]
      }
    },
...
    "/api/v1/security/passkeys/register/complete": {
      "post": {
        "operationId": "complete_passkey_registration_api_v1_security_passkeys_register_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyRegistrationCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation problem"
          }
        },
        "summary": "Complete Passkey Registration",
        "tags": [
          "security"
        ]
      }
    },
    "/api/v1/security/passkeys/{passkey_uuid}": {
      "delete": {
        "operationId": "delete_passkey_api_v1_security_passkeys__passkey_uuid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "passkey_uuid",
            "required": true,
            "schema": {
              "title": "Passkey Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation problem"
          }
        },
        "summary": "Delete Passkey",
        "tags": [
          "security"
        ]
      }
    },
    "/api/v1/security/pix": {
      "post": {
        "operationId": "update_pix_api_v1_security_pix_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PixUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PixUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation problem"
          }
        },
        "summary": "Update Pix",
        "tags": [
          "security"
        ]
      }
    },
...
    "/api/v1/security/totp/confirm": {
      "post": {
        "operationId": "confirm_totp_api_v1_security_totp_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TOTPConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation problem"
          }
        },
        "summary": "Confirm Totp",
        "tags": [
          "security"
        ]
      }
    },
    "/api/v1/security/totp/disable": {
      "post": {
        "operationId": "disable_totp_api_v1_security_totp_disable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TOTPDisableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation problem"
          }
        },
        "summary": "Disable Totp",
        "tags": [
          "security"
        ]
      }
    },
...
  }
}